[ethcore]: move client test types to test-helpers
