+++ ethcore/light/src/provider.rs
@@ -396,7 +396,7 @@ impl<L: AsLightClient> AsLightClient for LightProvider<L> {
 
 #[cfg(test)]
 mod tests {
-	use ethcore::client::{EachBlockWith, TestBlockChainClient};
+	use ethcore::test_helpers::{EachBlockWith, TestBlockChainClient};
 	use super::Provider;
 
 	#[test]
