+++ ethcore/src/json_tests/state.rs
@@ -18,8 +18,8 @@ use std::path::Path;
 use super::test_common::*;
 use pod::PodState;
 use trace;
-use client::{EvmTestClient, EvmTestError, TransactErr, TransactSuccess};
 use ethjson;
+use test_helpers::{EvmTestClient, EvmTestError, TransactErr, TransactSuccess};
 use types::transaction::SignedTransaction;
 use vm::EnvInfo;
 use super::SKIP_TEST_STATE;
