+++ ethcore/src/test_helpers/evm_test_client.rs
