+++ ethcore/src/test_helpers/mod.rs
@@ -16,6 +16,18 @@
 
 //! Set of different helpers for client tests
 
+mod test_client;
+mod evm_test_client;
+
+/// Re-export for tests only
+pub use evm::CreateContractAddress;
+/// Re-export for tests only
+pub use trie::TrieSpec;
+/// Re-export for tests only
+pub use self::test_client::{TestBlockChainClient, EachBlockWith, TestState};
+/// Re-export for tests only
+pub use self::evm_test_client::{EvmTestClient, EvmTestError, TransactErr, TransactSuccess};
+
 use std::path::Path;
 use std::sync::Arc;
 use std::{fs, io};
@@ -53,10 +65,6 @@ use spec::{Spec, self};
 use account_state::*;
 use state_db::StateDB;
 
-/// Re-export for tests only
-pub use evm::CreateContractAddress;
-/// Re-export for tests only
-pub use trie::TrieSpec;
 
 /// Creates test block with corresponding header
 pub fn create_test_block(header: &Header) -> Bytes {
