+++ ethcore/sync/src/chain/handler.rs
@@ -801,11 +801,11 @@ impl SyncHandler {
 #[cfg(test)]
 mod tests {
 	use client_traits::ChainInfo;
-	use ethcore::client::{EachBlockWith, TestBlockChainClient};
+	use ethcore::test_helpers::{EachBlockWith, TestBlockChainClient};
 	use parking_lot::RwLock;
 	use rlp::Rlp;
-	use std::collections::{VecDeque};
-	use tests::helpers::{TestIo};
+	use std::collections::VecDeque;
+	use tests::helpers::TestIo;
 	use tests::snapshot::TestSnapshotService;
 
 	use super::*;
