+++ ethcore/sync/src/chain/propagator.rs
@@ -336,7 +336,7 @@ impl SyncPropagator {
 #[cfg(test)]
 mod tests {
 	use client_traits::{BlockInfo, ChainInfo};
-	use ethcore::client::{EachBlockWith, TestBlockChainClient};
+	use ethcore::test_helpers::{EachBlockWith, TestBlockChainClient};
 	use parking_lot::RwLock;
 	use rlp::Rlp;
 	use std::collections::VecDeque;
