+++ rpc/src/v1/tests/helpers/miner_service.rs
@@ -23,8 +23,9 @@ use bytes::Bytes;
 use client_traits::{Nonce, StateClient};
 use engine::{Engine, signer::EngineSigner};
 use ethcore::block::SealedBlock;
-use ethcore::client::{PrepareOpenBlock, EngineInfo, TestState};
+use ethcore::client::{PrepareOpenBlock, EngineInfo};
 use ethcore::miner::{self, MinerService, AuthoringParams, FilterOptions};
+use ethcore::test_helpers::TestState;
 use ethereum_types::{H256, U256, Address};
 use miner::pool::local_transactions::Status as LocalTransactionStatus;
 use miner::pool::{verifier, VerifiedTransaction, QueueStatus};
