+++ rpc/src/v1/tests/mocked/eth.rs
@@ -21,7 +21,7 @@ use std::time::{Instant, Duration, SystemTime, UNIX_EPOCH};
 
 use accounts::AccountProvider;
 use client_traits::BlockChainClient;
-use ethcore::client::{EachBlockWith, TestBlockChainClient};
+use ethcore::test_helpers::{EachBlockWith, TestBlockChainClient};
 use ethcore::miner::{self, MinerService};
 use ethereum_types::{H160, H256, U256, Address, Bloom};
 use machine::executed::Executed;
