+++ rpc/src/v1/tests/mocked/personal.rs
@@ -20,7 +20,7 @@ use std::str::FromStr;
 use bytes::ToPretty;
 use accounts::AccountProvider;
 use ethereum_types::{Address, H520, U256};
-use ethcore::client::TestBlockChainClient;
+use ethcore::test_helpers::TestBlockChainClient;
 use jsonrpc_core::IoHandler;
 use parking_lot::Mutex;
 use types::transaction::{Action, Transaction};
