+++ rpc/src/v1/tests/mocked/traces.rs
@@ -19,7 +19,7 @@ use std::sync::Arc;
 use machine::executed::Executed;
 use trace::trace::{Action, Res, Call};
 use trace::LocalizedTrace;
-use ethcore::client::TestBlockChainClient;
+use ethcore::test_helpers::TestBlockChainClient;
 use ethereum_types::{Address, H256};
 
 use types::transaction::CallError;
